pyproject.toml
[project]
name = "mch-extract"
version = "0.1.3"
description = "A Python package for extracting meteorological data from MeteoSwiss OpenData API"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Martin Spoto", email = "martin.spoto98@gmail.com"}
]
keywords = ["meteorology", "weather", "data", "meteoswiss", "switzerland"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
dependencies = [
    "more-itertools>=10.7.0",
    "polars-lts-cpu>=1.31.0",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.8.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
mch-extract = "mchextract.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_GITHUB_USERNAME/mch-extract"
Repository = "https://github.com/YOUR_GITHUB_USERNAME/mch-extract"
Issues = "https://github.com/YOUR_GITHUB_USERNAME/mch-extract/issues"
Documentation = "https://github.com/YOUR_GITHUB_USERNAME/mch-extract#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mchextract"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[[tool.uv.index]]
explicit = true
name = "testpypi"
publish-url = "https://test.pypi.org/legacy/"
url = "https://test.pypi.org/simple/"
